+++ orb/api/resources.py
@@ -120,9 +120,6 @@ class ResourceResource(ModelResource):
             resource = Resource.objects.get(create_user=bundle.request.user,title =bundle.data['title'])
             rr = ResourceResource()
             bundle = rr.build_bundle(obj=resource,request=request)
-            print bundle
-            #return HttpResponse(#content=json.dumps({'object':'hello'}),
-            #                    content_type="application/json; charset=utf-8")
             raise BadRequest(_(u'You have already uploaded a resource with this title'))
         except Resource.DoesNotExist:
             pass